requirements.txt
fastapi
uvicorn[standard]
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pymysql
redis
requests
mysql-connector-python
python-dotenv
bilibili-api-python
pandas
jieba
numpy
uvloop
httptools
aiohttp 